refactor formatter to not do parsing
parsing doesn't depend on the environment and is the opposite of
formatting, so it's better in its own module. Also makes it
easier to reuse the parsing code in the desktop client.

moves:
* parseMailToUrl from MailUtils.js to MailAddressParser.js
* mail address parsing code from Formatter.js to MailAddressParser.js
* parseTime from Formatter.js to TimeParser.js
* unit tests to appropriate files

#2968
