+++ src/login/contactform/ContactFormRequestDialog.js
@@ -4,7 +4,7 @@ import {Dialog, DialogType} from "../../gui/base/Dialog"
 import {Button} from "../../gui/base/Button"
 import {Type} from "../../gui/base/TextFieldN"
 import {lang} from "../../misc/LanguageViewModel"
-import {formatStorageSize, getCleanedMailAddress} from "../../misc/Formatter"
+import {formatStorageSize} from "../../misc/Formatter"
 import {ConversationType, InputFieldType, Keys, MailMethod, MAX_ATTACHMENT_SIZE, PushServiceType} from "../../api/common/TutanotaConstants"
 import {animations, height} from "../../gui/animation/Animations"
 import {assertMainOrNode} from "../../api/common/Env"
@@ -36,6 +36,7 @@ import {TextFieldN} from "../../gui/base/TextFieldN"
 import {showProgressDialog} from "../../gui/dialogs/ProgressDialog"
 import {DropDownSelectorN} from "../../gui/base/DropDownSelectorN"
 import type {InputField} from "../../api/entities/tutanota/InputField"
+import {getCleanedMailAddress} from "../../misc/parsing/MailAddressParser"
 import {ofClass} from "../../api/common/utils/PromiseUtils"
 
 assertMainOrNode()
