+++ src/settings/IdentifierListViewer.js
@@ -10,7 +10,6 @@ import {createPushIdentifier, PushIdentifierTypeRef} from "../api/entities/sys/P
 import {logins} from "../api/main/LoginController"
 import {Icons} from "../gui/base/icons/Icons"
 import {PushServiceType} from "../api/common/TutanotaConstants"
-import {getCleanedMailAddress} from "../misc/Formatter"
 import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import {worker} from "../api/main/WorkerClient"
 import {Dialog} from "../gui/base/Dialog"
@@ -27,6 +26,7 @@ import type {EntityUpdateData} from "../api/main/EventController"
 import {isUpdateForTypeRef} from "../api/main/EventController"
 import type {User} from "../api/entities/sys/User"
 import {showNotAvailableForFreeDialog} from "../misc/SubscriptionDialogs"
+import {getCleanedMailAddress} from "../misc/parsing/MailAddressParser"
 import type {TranslationKey} from "../misc/LanguageViewModel"
 import {ofClass} from "../api/common/utils/PromiseUtils"
