+++ test/client/Suite.js
@@ -42,6 +42,8 @@ import "./misc/ParserTest"
 import "./templates/TemplateEditorModelTest"
 import "./misc/SchedulerTest"
 import "./subscription/PriceUtilsTest"
+import "./misc/parsing/MailAddressParserTest"
+import "./misc/FormatValidatorTest"
 import o from "ospec"
 import {random} from "../../src/api/worker/crypto/Randomizer"
 import {EntropySrc} from "../../src/api/common/TutanotaConstants"
